contracts/requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
bitarray==3.7.1
cached-property==2.0.1
cbor2==5.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
click==8.3.0
colorama==0.4.6
coverage==7.10.7
cytoolz==1.0.1
eth-account==0.13.7
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-stdlib==0.2.8
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
ghp-import==2.1.0
hexbytes==1.3.1
hypothesis==6.140.2
idna==3.10
immutables==0.21
iniconfig==2.1.0
Jinja2==3.1.6
lark==1.3.0
lru-dict==1.3.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.41
mkdocs-material-extensions==1.3.1
packaging==24.2
paginate==0.5.7
parsimonious==0.10.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
py-ecc==8.0.0
py-evm==0.12.1b1
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pymdown-extensions==10.16.1
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rlp==4.1.0
six==1.17.0
snekmate @ git+https://github.com/pcaversaccio/snekmate.git@307fc620000f70241a626d7ef4ec9c3183f26546
sortedcontainers==2.4.0
titanoboa==0.2.7
toolz==1.0.0
trie==3.1.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
vvm==0.3.2
vyper==0.4.3
watchdog==6.0.0
wheel==0.45.1
